+++ api/opentrons/drivers/smoothie_drivers/__init__.py
@@ -1,0 +1,10 @@
+class SmoothieDriver(object):
+
+    def __init__(self):
+        pass
+
+
+class VirtualSmoothie(object):
+
+    def __init__(self):
+        pass
