+++ api/opentrons/drivers/smoothie_drivers/v1_2_0/driver.py
@@ -1,5 +1,3 @@
-import configparser
-import glob
 import json
 import math
 import os
@@ -12,7 +10,7 @@ import serial
 
 from opentrons.util.log import get_logger
 from opentrons.util.vector import Vector
-from opentrons.drivers.connection import VirtualSmoothie, SmoothieDriver
+from opentrons.drivers.smoothie_drivers import VirtualSmoothie, SmoothieDriver
 
 from opentrons.util import trace
 
@@ -135,41 +133,6 @@ class SmoothieDriver_1_2_0(SmoothieDriver):
         if not self.ot_version:
             self.get_ot_version()
         return self.ot_one_dimensions[self.ot_version]
-
-    def get_serial_ports_list(self):
-        """ Lists serial port names
-
-            :raises EnvironmentError:
-                On unsupported or unknown platforms
-            :returns:
-                A list of the serial ports available on the system
-        """
-        if sys.platform.startswith('win'):
-            ports = ['COM%s' % (i + 1) for i in range(256)]
-        elif (sys.platform.startswith('linux') or
-              sys.platform.startswith('cygwin')):
-            # this excludes your current terminal "/dev/tty"
-            ports = glob.glob('/dev/tty*')
-            # ignore Smoothie's local storage if linux (temporary work-around)
-            self.ignore_smoothie_sd = True
-        elif sys.platform.startswith('darwin'):
-            ports = glob.glob('/dev/tty.*')
-        else:
-            raise EnvironmentError('Unsupported platform')
-
-        result = []
-        port_filter = {'usbmodem', 'COM', 'ACM', 'USB'}
-        for port in ports:
-            try:
-                if any([f in port for f in port_filter]):
-                    s = serial.Serial(port)
-                    s.close()
-                    result.append(port)
-            except Exception as e:
-                log.debug(
-                    'Exception in testing port {}'.format(port))
-                log.debug(e)
-        return result
 
     def disconnect(self):
         if self.is_connected() and self.connection:
