+++ api/opentrons/drivers/smoothie_drivers/v2_0_0/virtual_smoothie.py
@@ -1,6 +1,6 @@
 import re
 
-from opentrons.drivers.connection import VirtualSmoothie
+from opentrons.drivers.smoothie_drivers import VirtualSmoothie
 from opentrons.util import log
 
 log = log.get_logger(__name__)
